containers/predictive_maintenance/requirements.txt
certifi==2023.5.7
influxdb-client==1.36.1
joblib==1.2.0
lightgbm==3.3.5
numpy==1.25.0
paho-mqtt==1.6.1
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.3
reactivex==4.0.4
scikit-learn==1.2.2
scipy==1.11.0
six==1.16.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
